requirements.txt

#basic libraries
asgiref==3.8.1
Django==4.2
django-filter==25.1
Markdown==3.7
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
tzdata==2025.1

#api libraries
djangorestframework==3.15.2

#third party libraries
django-filter==25.1

